Combine AE2 module and defaultrecipe classes.
